#23 Added serialization tests for the stream entities
