--- src/main/scala/com/danielasfregola/twitter4s/entities/streaming/Envelop.scala
@@ -1,5 +1,0 @@
-package com.danielasfregola.twitter4s.entities.streaming
-
-case class Envelop(for_user: Long, message: StreamingMessage) extends StreamingMessage
-
-case class EnvelopStringified(for_user: String, message: StreamingMessage) extends StreamingMessage
