+++ src/main/scala/com/danielasfregola/twitter4s/http/clients/StreamingOAuthClient.scala
@@ -51,12 +51,12 @@ private[twitter4s] trait StreamingOAuthClient extends OAuthClient {
       json4sUnmarshaller[DisconnectMessage].asInstanceOf[Unmarshaller[StreamingMessage]],
       json4sUnmarshaller[WarningMessage].asInstanceOf[Unmarshaller[StreamingMessage]],
       json4sUnmarshaller[Event].asInstanceOf[Unmarshaller[StreamingMessage]],
+      json4sUnmarshaller[ControlMessage].asInstanceOf[Unmarshaller[StreamingMessage]],
       json4sUnmarshaller[FriendsLists].asInstanceOf[Unmarshaller[StreamingMessage]],
       json4sUnmarshaller[FriendsListsStringified].asInstanceOf[Unmarshaller[StreamingMessage]],
       json4sUnmarshaller[DirectMessage].asInstanceOf[Unmarshaller[StreamingMessage]],
       json4sUnmarshaller[UserEnvelop].asInstanceOf[Unmarshaller[StreamingMessage]],
-      json4sUnmarshaller[UserEnvelopStringified].asInstanceOf[Unmarshaller[StreamingMessage]],
-      json4sUnmarshaller[ControlMessage].asInstanceOf[Unmarshaller[StreamingMessage]]
+      json4sUnmarshaller[UserEnvelopStringified].asInstanceOf[Unmarshaller[StreamingMessage]]
     )
 
   implicit val eventTargetObjectUnmarshaller: Unmarshaller[EventTargetObject] =
