+++ src/test/resources/twitter/streaming/public_control_messages.json
@@ -1,0 +1,7 @@
+[
+  {
+    "control": {
+      "control_uri": "/1.1/site/c/01_225167_334389048B872A533002B34D73F8C29FD09EFC50"
+    }
+  }
+]
