+++ src/test/resources/twitter/streaming/public_friends_lists.json
@@ -1,0 +1,5 @@
+[
+  {
+    "friends": [0,1,2,3,4]
+  }
+]
