+++ src/test/resources/twitter/streaming/public_user_envelops.json
@@ -1,0 +1,8 @@
+[
+  {
+    "for_user": 1888,
+    "message": {
+      "friends": [0, 1, 2, 3, 4]
+    }
+  }
+]
