+++ src/test/scala/com/danielasfregola/twitter4s/http/clients/streaming/statuses/TwitterStatusClientSpec.scala
@@ -1,13 +1,9 @@
 package com.danielasfregola.twitter4s.http.clients.streaming.statuses
 
-import com.danielasfregola.twitter4s.entities.streaming._
-import com.danielasfregola.twitter4s.entities.Tweet
 import com.danielasfregola.twitter4s.http.clients.streaming.TwitterStreamingSpecContext
 import com.danielasfregola.twitter4s.util.ClientSpec
 import spray.http.Uri.Query
 import spray.http.{HttpEntity, _}
-
-import scala.io.Source
 
 class TwitterStatusClientSpec extends ClientSpec {
 
@@ -47,27 +43,6 @@ class TwitterStatusClientSpec extends ClientSpec {
             request.uri.query === Query("stall_warnings=false")
         }.respondWithOk.await
       result.isInstanceOf[Unit] should beTrue
-    }
-
-    "parse a stream of data and output the corresponding entities" in new TwitterStatusClientSpecContext {
-      val processor = system.actorOf(StreamingActor.props(self))
-
-      Source.fromURL(getClass.getResource("/fixtures/streaming/public.json")).getLines.foreach { line =>
-        processor ! MessageChunk(HttpData(line) +: HttpData("\r\n"))
-      }
-
-      val messages: Seq[StreamingUpdate] =
-        (loadJsonAs[Seq[Tweet]]("/twitter/streaming/public_tweets.json") ++
-         loadJsonAs[Seq[LimitNotice]]("/twitter/streaming/public_limit_notices.json") ++
-         loadJsonAs[Seq[DisconnectMessage]]("/twitter/streaming/public_disconnected_messages.json") ++
-         loadJsonAs[Seq[LocationDeletionNotice]]("/twitter/streaming/public_location_deletion_notices.json") ++
-         loadJsonAs[Seq[StatusWithheldNotice]]("/twitter/streaming/public_status_withheld_notices.json") ++
-         loadJsonAs[Seq[UserWithheldNotice]]("/twitter/streaming/public_user_withheld_notices.json") ++
-         loadJsonAs[Seq[WarningMessage]]("/twitter/streaming/public_warning_messages.json") ++
-         loadJsonAs[Seq[StatusDeletionNotice]]("/twitter/streaming/public_status_deletion_notices.json"))
-        .map (StreamingUpdate(_))
-
-      expectMsgAllOf(messages: _*)
     }
   }
 }
